moved Rhyme impelmentation into CommonRhymeBuilderImpl
